Refactor util.py
